requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.8.2
